created storage-image file in middlewares folder
